# docs/openapi-gateway.yaml
openapi: 3.1.0
info:
  title: snitchr Gateway API
  version: 0.1.0
  description: |
    API for the snitchr monitoring gateway. Runs on Cloudflare Workers, Node.js, or Bun.

    Two auth models:
    - **Session cookie** — HMAC-signed expiry timestamp in httpOnly cookie (24h TTL). Used by all dashboard/admin endpoints.
    - **Bearer token** — per-machine 48-char hex secret. Used by the agent ingest endpoint only.
  license:
    name: MIT
    url: https://github.com/thinking-tools/snitchr/blob/main/LICENSE
  contact:
    name: snitchr
    url: https://snitchr.sh
servers:
  - url: https://{host}
    description: snitchr gateway
    variables:
      host:
        default: localhost:8787
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: HMAC-signed expiry timestamp (24h TTL)
    bearerToken:
      type: http
      scheme: bearer
      description: Per-machine 48-char hex secret

  # ── Reusable schemas ──────────────────────────────────────────────────
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
    Ok:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
    EventLevel:
      type: integer
      enum: [0, 1, 2, 3, 4, 5]
      description: |
        0=HEARTBEAT, 1=INFO, 2=WARN, 3=ERROR, 4=CRITICAL, 5=SHUTDOWN
...
    StorageMode:
      type: string
      enum: [r2, s3, filesystem, memory]
...
    AgentFeatures:
      type: object
      properties:
        authWatch:
          type: boolean
        fileIntegrity:
          type: boolean
        processWatch:
          type: boolean
        portWatch:
          type: boolean
        mountWatch:
          type: boolean
        cpuAlerts:
          type: boolean
        memAlerts:
          type: boolean
        diskAlerts:
          type: boolean
    NotificationChannel:
      oneOf:
        - type: object
          required: [type, endpoint, enabled]
          additionalProperties: false
          properties:
            type:
              type: string
              enum: [ntfy]
            endpoint:
              type: string
              format: uri
            enabled:
              type: boolean
        - type: object
          required: [type, url, enabled]
          additionalProperties: false
          properties:
            type:
              type: string
              enum: [webhook]
            url:
              type: string
              format: uri
            secret:
              type: string
            enabled:
              type: boolean
        - type: object
          required: [type, webhookUrl, enabled]
          additionalProperties: false
          properties:
            type:
              type: string
              enum: [slack]
            webhookUrl:
              type: string
              format: uri
              pattern: '^https://hooks\.slack\.com/'
            enabled:
              type: boolean
        - type: object
          required: [type, enabled]
          additionalProperties: false
          properties:
            type:
              type: string
              enum: [webpush]
            enabled:
              type: boolean
      discriminator:
        propertyName: type
...
    WebPushSubscription:
      type: object
      required: [endpoint, keys]
      properties:
        endpoint:
          type: string
          format: uri
        expirationTime:
          type: ['number', 'null']
        keys:
          type: object
          required: [p256dh, auth]
          properties:
            p256dh:
              type: string
              minLength: 1
              description: Diffie-Hellman public key (base64url)
            auth:
              type: string
              minLength: 1
              description: Authentication secret (base64url)
  responses:
    Unauthorized:
      description: Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

# ── Paths ─────────────────────────────────────────────────────────────
paths:
...
  /api/config:
    get:
      tags: [Setup & Auth]
      summary: Get dashboard config
      description: Returns machine IDs, alert threshold, and storage mode.
      operationId: getConfig
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Dashboard config
          content:
            application/json:
              schema:
                type: object
                required: [machineIds, alertThreshold, storageMode]
                properties:
                  machineIds:
                    type: array
                    items:
                      type: string
                      format: uuid
                  alertThreshold:
                    $ref: '#/components/schemas/EventLevel'
                  storageMode:
                    $ref: '#/components/schemas/StorageMode'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Agent Ingest ────────────────────────────────────────────────────
...
  /api/settings:
    get:
      tags: [Settings]
      summary: Get all settings
      description: Returns alert threshold, heartbeat timeout, notification channels, agent features, storage mode, and machine count.
      operationId: getSettings
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema:
                type: object
                required: [alertThreshold, heartbeatTimeout, notifications, agentFeatures, storageMode, machineCount]
                properties:
                  alertThreshold:
                    type: integer
                    minimum: 1
                    maximum: 4
                  heartbeatTimeout:
                    type: integer
                    minimum: 30
                    maximum: 3600
                  notifications:
                    type: object
                    additionalProperties:
                      $ref: '#/components/schemas/NotificationChannel'
                  agentFeatures:
                    $ref: '#/components/schemas/AgentFeatures'
                  storageMode:
                    $ref: '#/components/schemas/StorageMode'
                  machineCount:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags: [Settings]
      summary: Update settings
      description: Partially update monitoring thresholds, notification channels, or agent feature flags.
      operationId: updateSettings
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                alertThreshold:
                  type: integer
                  minimum: 1
                  maximum: 4
                heartbeatTimeout:
                  type: integer
                  minimum: 30
                  maximum: 3600
                notifications:
                  type: object
                  additionalProperties:
                    $ref: '#/components/schemas/NotificationChannel'
                agentFeatures:
                  $ref: '#/components/schemas/AgentFeatures'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/settings/password:
    post:
      tags: [Settings]
      summary: Change admin password
      description: Verify current password and set a new one. Issues a new session cookie.
      operationId: changePassword
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword, confirmPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
                confirmPassword:
                  type: string
      responses:
        '200':
          description: Password changed; new session cookie set
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/settings/test-notification:
    post:
      tags: [Settings]
      summary: Test a notification channel
      description: Send a test notification to the specified channel.
      operationId: testNotification
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel]
              properties:
                channel:
                  type: string
                  minLength: 1
                  description: Notification channel key
      responses:
        '200':
          description: Test notification sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/settings/reset:
    post:
      tags: [Settings]
      summary: Factory reset
      description: Wipe all data and config. Requires current password for verification. Clears session and redirects to setup.
      operationId: factoryReset
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        '200':
          description: Reset complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Push Notifications ──────────────────────────────────────────────
  /api/push/vapid-key:
    get:
      tags: [Push Notifications]
      summary: Get VAPID public key
      description: Returns the VAPID public key for web push subscriptions. Auto-generates keys on first call.
      operationId: getVapidKey
      security:
        - sessionCookie: []
      responses:
        '200':
          description: VAPID key
          content:
            application/json:
              schema:
                type: object
                required: [publicKey]
                properties:
                  publicKey:
                    type: string
                    description: VAPID public key (base64url)
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/push/subscribe:
    post:
      tags: [Push Notifications]
      summary: Subscribe to web push
      description: Register a browser for push notifications. Auto-enables webpush channel. Max 50 subscriptions (deduped by endpoint).
      operationId: subscribePush
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebPushSubscription'
      responses:
        '200':
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/push/unsubscribe:
    post:
      tags: [Push Notifications]
      summary: Unsubscribe from web push
      description: Remove a browser subscription. Auto-disables webpush channel if no subscriptions remain.
      operationId: unsubscribePush
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint]
              properties:
                endpoint:
                  type: string
                  format: uri
                  minLength: 1
      responses:
        '200':
          description: Unsubscribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── System ──────────────────────────────────────────────────────────
...